requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
awesome-slugify==1.6.5
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
daphne==4.0.0
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.2
django-allauth==0.51.0
django-appconf==1.0.5
django-ckeditor==6.5.1
django-filter==22.1
django-imagekit==4.1.0
django-js-asset==2.0.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql==0.0.3
mysqlclient==2.1.1
oauthlib==3.2.2
pilkit==2.0
Pillow==9.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.5
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.5
Unidecode==0.4.21
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
zope.interface==5.5.0
